Moved MipsMachine.java into machine subpackage
