+++ src/main/java/io/github/danielt3131/mipsemu/machine/MipsMachine.java
@@ -1,4 +1,4 @@
-package io.github.danielt3131.mipsemu;
+package io.github.danielt3131.mipsemu.machine;
 
 public class MipsMachine {
